+++ tools/pd-simulator/simulator/info/stats.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dto
+package info
 
 import "github.com/pingcap/kvproto/pkg/pdpb"
