+++ tools/pd-simulator/simulator/node.go
@@ -25,7 +25,7 @@ import (
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	"github.com/pingcap/pd/tools/pd-simulator/simulator/cases"
-	"github.com/pingcap/pd/tools/pd-simulator/simulator/dto"
+	"github.com/pingcap/pd/tools/pd-simulator/simulator/info"
 	"github.com/pingcap/pd/tools/pd-simulator/simulator/simutil"
 )
 
@@ -39,7 +39,7 @@ const (
 type Node struct {
 	*metapb.Store
 	sync.RWMutex
-	stats                    *dto.StoreStats
+	stats                    *info.StoreStats
 	tick                     uint64
 	wg                       sync.WaitGroup
 	tasks                    map[uint64]Task
@@ -62,7 +62,7 @@ func NewNode(s *cases.Store, pdAddr string, ioRate int64) (*Node, error) {
 		Labels:  s.Labels,
 		State:   s.Status,
 	}
-	stats := &dto.StoreStats{
+	stats := &info.StoreStats{
 		StoreStats: pdpb.StoreStats{
 			StoreId:   s.ID,
 			Capacity:  s.Capacity,
